backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.1.0
tensorflow==2.12.0
deepface==0.0.79
pillow==11.2.1
opencv-python-headless==4.8.0.74
